Add mixed domain tests for EAP 6.4.0
