+++ testsuite/mixed-domain/src/test/java/org/jboss/as/test/integration/domain/mixed/Version.java
@@ -42,7 +42,8 @@ public @interface Version {
 
     enum AsVersion {
         EAP_6_2_0(EAP, "6.2.0"),
-        EAP_6_3_0(EAP, "6.3.0");
+        EAP_6_3_0(EAP, "6.3.0"),
+        EAP_6_4_0(EAP, "6.4.0");
 
         final String basename;
         final String version;
